Rename gen_compare to gen_equality
These tests check the equality functionality. In preparation of adding
a Sort ability (which defines a compare method), naming the quality
tests `gen_compare` creates ambiguity, so this renames them.
