+++ crates/compiler/test_gen/src/tests.rs
@@ -5,9 +5,9 @@
 #![allow(clippy::float_cmp)]
 
 pub mod gen_abilities;
-pub mod gen_compare;
 pub mod gen_definitions;
 pub mod gen_dict;
+pub mod gen_equality;
 pub mod gen_erased;
 pub mod gen_list;
 pub mod gen_num;
